Add in read in of data from stupid structure
Wow re.split is a neat function. I feel like a doorknob for never using it.

just it iterate through the list of datapoints for the positive and negative
sentiment word lists of the 50k entries. Leaving out the unsup version right now
since this doesn't make sense to use.